hw5/notification_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
aio-pika
pytest
pytest-asyncio
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
prometheus-fastapi-instrumentator
tenacity
python-logstash-async
